Silence compiler warning

+++ test/segmented_pi_table.cpp
@@ -37,7 +37,7 @@ int main()
 
   int64_t limit = dist(gen);
   int64_t segment_size = iroot<3>(limit);
-  int64_t threads = 1;
+  int threads = 1;
 
   PiTable pi(limit);
   SegmentedPiTable segmentedPi(limit, segment_size, threads);
